UsingCluster_NN/requirements.txt
torch
torchvision
scikit-learn
pyyaml